+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+* Issue **#13** : Added value functions `true()`, `false()`, `null()` and `err()`.
+
 * Issue **#14** : Added `match` boolean function.
 
 * Issue **#12** : Added `variance` and `stDev` functions.